fixed autoOpen and snipes (before grace period ends)

+++ src/main/java/com/max2341/fruitbin/AuctionInfo.java
@@ -15,7 +15,7 @@ public class AuctionInfo {
 	 }
 
 	AuctionItem auction;
-	int profit;
+	long profit;
 	int profitPercent;
 	float price;
 	float lowestBin;
